rustdoc: use the same URL escape rules for fragments as for examples
